Update SLI alert to page below 95% for 5m
The original alert was set to "max < 0.95 over 5 minutes," and that was
too noisy. We switched it to 10 minutes, then changed it to "at all
times," then to "at all times over the past 30 minutes," which was not
sensitive enough.

I'm bumping it up to "at all times during the past 5
minutes," which should mean that we only get notified if the Java SLI
app is below 95% for 5 straight minutes.

[#151486134] **Investigate** why SLI failures didn't page us for a while during the S3 outage on Sep 14