Diego Free RAM is monitored absolutely, not as a percent
Alerts on less than 700GB free (previously alerted <5% free)

We were grossly over-provisioned; we had 4.4 TB RAM free, and that's
too much. We needed an absolute minimum of 281 GB RAM free, and that's
to accommodate the evacuation of 6 full Diego cells simultaneously
(BOSH's max_in_flight is 6). We have decided that 960 GB is a reasonable
amount of free RAM. This reduction drops the number of r4.xlarge
instances by 70, which should save close to $20k/month.

[#158696261] Save SLI Dashboard & Diego memory monitor for Avail Memory Conditional Formatting

Signed-off-by: Jesse Malone <jmalone@pivotal.io>