[Backend TLS invalid cert] improve monitor
Since the 'backend_invalid_tls_cert' metric is a lifetime counter,
alerting off of the the value of this metric is not as useful as
measuring the rate at which this metric changes. A threshold of 10 is
still a bit arbitrary but should provide an "action-worthy" signal to
look into an issue within the platform.

[#167662632]