Update diego monitor to cumulatively alert on GardenHealthCheckFailed
After we fixed this monitor to use a non-deprecated metric, we noticed
that it is sensitive to situations when we scale down diego cells. We
have updated this monitor so that it will not alert on "no data" for
individual cells but rather only alert when the cumulative health of the
cells is triggered.

[#168036184]