FIX: remove deployment filter for GardenHealthCheckFailed monitor
Since we want to know about all of the reps' status, we can remove the
filter for a specific deployment. This makes things easier for
monitoring PWS as we have many deployments with reps. An alternative
approach is to add multiple deployment filters but that will involve a
bit more work.