Remove alert templates for PWS Platform-owned monitors
The monitors that these templates create will no longer be used because
we (the PWS Platform team) have switched to using Wavefront for
monitoring and alerting. Although this repo was, in the past, intended
for consumption by other teams, we no longer support it as a product and
discourage its use.

We calculated the list of templates to delete by searching for the
values in the "Name" column in the [monitor migration
spreadsheet](https://docs.google.com/spreadsheets/d/1Ezr-usH8wkU-0XnWDOjD5E0PBJg6i54FKlGVM8MdF80/edit#gid=2030347806)
for rows that had "PWS Platform" in the "Team" column. For names that
started with "prod", we removed the "prod" portion prior to searching
because that portion of the name is interpolated by the config tooling;
in the templates themselves it shows up as '<%= environment %>`.

[#170580001] Delete (or disable) PWS Platform-owned monitors in Datadog

Co-authored-by: Tom Viehman <tviehman@pivotal.io>